requirements.txt
torch
librosa
pyyaml
scipy
argparse
wandb
